requirements.txt
asgiref==3.5.2
django-autotranslate==1.1.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-crispy-forms==1.14.0
django-etc==1.4.0
django-hitcount==1.3.5
django-linguist==0.6.0
django-import-export==3.0.1
django-rosetta==0.9.8
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
